feat(cli): add a `--no-remote` flag to lock the application on the cli-invoked channel (#455)
This will disable the remote control panel and associated actions
entirely. This is useful when the remote control is not needed or when
the user wants `tv` to run in single-channel mode (e.g. when using it as
a file picker for a script or embedding it in a larger application).