feat(cli): add a `--no-help` flag to allow disabling showing the help panel
This will disable the help panel and associated toggling actions
entirely. This is useful when the help panel is not needed or
when the user wants tv to run with a minimal interface (e.g. when
using it as a file picker for a script or embedding it in a larger
application).