fix(github): copy github prototypes' content directly
The current implementation downloads the raw prototype files from
github, then deserializes them into `ChannelPrototype` objects and
re-serializes them to the local config.
Because TOML can express same things in multiple ways, it isn't trivial
to have consistent TOML formatting results when re-serializing the
files.
I consider this to be a bug, which is why I'm now directly copying the
downloaded file content to the local config, while keeping a sanity
check by attempting to deserialize it along the way.