feat(actions): improve actions behavior and add support for `exec` (#689)
## 📺 PR Description

This improves support for `fork` and adds the logic for the `exec`
feature.

I'm dropping the `OutputMode` logic for now as it doesn't seem that
useful and kind of gets in the way right now. We might add that back
later on.

## Checklist

<!-- a quick pass through the following items to make sure you haven't
forgotten anything -->

- [ ] my commits **and PR title** follow the [conventional
commits](https://www.conventionalcommits.org/en/v1.0.0/#summary) format
- [ ] if this is a new feature, I have added tests to consolidate the
feature and prevent regressions
- [ ] if this is a bug fix, I have added a test that reproduces the bug
(if applicable)
- [ ] I have added a reasonable amount of documentation to the code
where appropriate