test: support passing custom tv binary path via env `TV_BIN_PATH` (#715)
## 📺 PR Description

Add a compile time env var override for the path to the `tv` binary. The
tests that run `tv` expect the path to be `./target/debug/tv`, but if
you pass a target to cargo, or if you run tests with `--release`, that
path doesn't work anymore.

This came up for the FreeBSD people (see
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=288915#c5), and also
for me while updating the Chimera Linux package.

For an example of how this can be used, see
https://github.com/jcgruenhage/cports/blob/a1dd86b5d24de91f1f2b72afa4c371753d074e1d/user/television/template.py#L5-L7.

## Checklist

<!-- a quick pass through the following items to make sure you haven't
forgotten anything -->

- [x] my commits **and PR title** follow the [conventional
commits](https://www.conventionalcommits.org/en/v1.0.0/#summary) format
- [x] if this is a new feature, I have added tests to consolidate the
feature and prevent regressions
- [x] if this is a bug fix, I have added a test that reproduces the bug
(if applicable)
- [x] I have added a reasonable amount of documentation to the code
where appropriate

Co-authored-by: Alex Pasmantier <47638216+alexpasmantier@users.noreply.github.com>