fix(cable): fix alias channel cli invocation sometimes going to bg
fixes #735

The problem seems to lie in interactive mode shells trying to reach the
tty before tv initializes it resulting in a SIGTTOU.
The fix simply delays that with a subshell which is kind of a hack but
does seem to work consistently.
Will need a more robust sync mechanism in the future.